requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.2.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-timezone-field==6.0.1
h11==0.14.0
humanize==4.8.0
idna==3.4
kombu==5.3.2
lxml==4.9.3
numpy==1.26.0
outcome==1.2.0
pandas==2.1.1
prompt-toolkit==3.0.39
psycopg2==2.9.9
pycparser==2.21
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tornado==6.3.3
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
wsproto==1.2.0
